# Conflicts:
#	bruno-api-pg/pedido/base/add_pedido.bru
#	bruno-api-pg/pedido/residuo/add_residuo.bru
#	bruno-api-pg/pedido/residuo/get_residuos.bru
#	src/main/java/org/purpura/apipg/controller/pedido/base/PedidoResiduoController.java

+++ src/main/java/org/purpura/apipg/dto/schemas/pedido/base/PedidoResiduoResponseDTO.java
@@ -28,7 +28,7 @@ public class PedidoResiduoResponseDTO {
     private Double preco;
 
     @Schema(description = "Quantidade do item", example = "10")
-    private Integer quantidade;
+    private Long quantidade;
 
     @Schema(description = "Peso do item", example = "6.00")
     private Double peso;
